Generate bulk request body as string
With this commit we generate the bulk request body as string instead of
as dict. This avoids encoding JSON on the fly.

We also clean up problematic escape sequences in data files as they
would trip Elasticsearch because we don't run these strings through a
JSON encoder anymore but just copy them verbatim. While we could replace
these values on the fly, we don't want to do this on the hot code path.