Update Random event regarding pipeline usage
Now that are going to support pipelines there is no need to pregenerate the
user_agent or the geoip field when the pipeline is defined.

Also due to the need to drop the field for the pipeline we updated how we
generate the json event in order to allow easier refactoring.