Run python unit tests for Randomevent
Since we changed how we generate the json events there was the need to reformat
the json structure in a more safer way.
This adds the ability to run python tests against the python code.

To execute the tests you can use this command from the root directory
```
python3.7 -m unittest \
          discover \
          --locals \
          --start-directory "." \
          --pattern "*test*.py"
```