Force merge index for frozen benchmarks
With this commit we force-merge the index that is used in the frozen
challenge to a single segment. We also set the refresh interval to its
default value to ensure an optimization in Elasticsearch can kick in
that disables refreshes if no searches are performed. This should
improve indexing speed a bit (we don't care about indexing speed in that
challenge as this only there to generate a decently sized corpus for the
query benchmark).