Remove custom create index runner
With this commit we remove the custom runner `createindex` which creates
index templates, matching indices and also aliases and replace it with
runners that are available out of the box with Rally. Furthermore, we
cleanup some duplication and have removed the challenge
`append-no-conflicts` which acted as a dummy default challenge but does
not serve any other purpose.

Relates #36