Simplify timestamp generator
With this commit we materialize relative timestamps eagerly so we can
easily generate timestamps on the hot code path without making any
further decisions. This slightly changes the semantics of the start
timestamp of the timestamp generator (it's now evaluated when the object
is created instead of when its generator method is first invoked) but
this does not matter in practice as both calls happen within a very
short time period.