Allow to limit the daily logging volume
With this commit we add a new parameter `daily_logging_volume` that
allows users to limit the amount of (raw) data that is generated per
(logical) day. If the limit is hit, we simulate that the next document
is generated on the next day. We also introduce the parameter
`number_of_days` that can be used to determine for how many days logs
should be generated.