Add challenge to index fixed amount of daily logs
With this commit we add a new challenge `index-logs-fixed-daily-volume` which
allows to ingest a fixed (but parameterizable) amount of raw logs per day. Also,
the number of days can be specified.

We also add a new parameter `daily_logging_volume` that allows users to limit
the amount of (raw) data that is generated per (logical) day. If the limit is
hit, we simulate that the next document is generated on the next day. With
another new parameter `number_of_days` we can configure for how many days logs
should be generated.

We also remove support to specify an `end_point` in the timestamp generator. 
This functionality was only used in one place and has also generated an
unrealistic distribution of timestamps (random across the whole range) and
should instead be simulated with an acceleration factor.

Finally, we materialize relative timestamps eagerly so we can easily
generate timestamps on the hot code path without making any further decisions.
This slightly changes the semantics of the start timestamp of the timestamp
generator (it's now evaluated when the object is created instead of when its
generator method is first invoked) but this does not matter in practice as both
calls happen within a very short time period.

Relates #38