Add progress indication for daily logging volume
With this commit we implement the property `pecent_completed` on the
Elasticlogs bulk source to provide an informational progress indication
for the challenge `index-logs-fixed-daily-volume`. Because we cannot
specify a total number of iterations or a specific time-period we need
to rely on feedback from the generator when all data has been ingested
and thus we also need to rely on the generator to indicate progress.