Add smoke test script
With this commit we add a smoke test script that allows to run a
benchmark in test mode against (almost) all challenges in this track. A
few challenges have been excluded intentionally because they rely on
other challenges being run first. While it would be possible to make
this work with workarounds we should wait for a proper solution with
elastic/rally#697