Use index template for update challenge
With this commit we use the proper index template for the bulk-update
challenge. Previously neither the index template nor the custom index
settings have been applied, resulting in a default index being created
unintentionally. This issue has been discovered by the smoke test script
that we have added in #47

Relates #50