Remove obsolete create-index tasks (#65)
This track defines an index template that defines common settings. When
bulk-indexing, indices will be automatically created based on this index
template. Therefore, explicit invocation of the create index API are
obsolete (unless they contain additional instructions such as creating
an alias).

With this commit we remove all these tasks and also related tasks (e.g.
it is not necessary to check for cluster health based on a specific
index if that index is not present yet).