Query more data for daily indices
With this commit we first fix the index pattern so queries can actually
match data. We also change the queries so we don't run any redundant
ones and modify them to retrieve more data. Finally, we modify the
Kibana runner so it stores some metadata about the query results (number
of total hits and the maximum value of `took` across all msearch
responses).